+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-
 =======   ========== ===========================================================================
 version   date       comment
 =======   ========== ===========================================================================
+?         ?          open_cache -> open_cache_folder
+                     logging level is not "info"
+                     added the possiblility to specify a text file containing primers 
+                     added the possiblility to specify a path to the ApE plasmid editor 
+                     (http://biologylabs.utah.edu/jorgensen/wayned/ape/)
+
+
+1.0.1     2016-03-10 Bugfix: for errors in IPython import if IPython is too old (<4.0)
+                     Bugfix: Large genbank records were not downloaded completely.
+1.0.0     -          Gel simulation added
 0.9.3     2015-06-03 Shelve does not work under MacOS under certain conditions.
                      This release tries to solve this by not specifying file extensions
                      for the cache files. Two functions are added, pydna.